+++ kotlin-sqlobject/src/main/kotlin/org/jdbi/v3/sqlobject/kotlin/RegisterKotlinMappersImpl.kt
@@ -15,7 +15,6 @@ package org.jdbi.v3.sqlobject.kotlin
 
 import org.jdbi.v3.core.config.ConfigRegistry
 import org.jdbi.v3.sqlobject.config.Configurer
-import org.jdbi.v3.sqlobject.config.RegisterKotlinMappers
 import java.lang.reflect.Method
 
 class RegisterKotlinMappersImpl : Configurer {