+++ RELEASE_NOTES.md
@@ -1,6 +1,7 @@
 # 3.14.0
   - immutables: support getter with @ColumnName, #1704
   - postgres: simple CRUD support for LargeObject API
+  - kotlin-sqlobject: fix package declaration of RegisterKotlinMappers
 
 # 3.13.0
   - Kotlin: respect default values in methods when passed null, #1690